requirements.txt
opencv-python
scikit-image
scikit-learn
numpy
joblib
Pillow
tk